+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.0.51] - 2025-12-12
+
 - Added option to use closest configuration file [[#602](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/602)]
+- Fix duplicate --config option issue [[#631](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/631)]
+- Fix freeze on save by executing apply in a background thread [[#647](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/647)]
+- Improve Python language lookup in Ruff integration [[#648](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/648)]
+- Enable Ruff format by default [[#646](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/646)]
+- Enhance Python SDK detection for Ruff in CLion [[#645](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/645)]
+- Add notification for native Ruff support in PyCharm 2025.3 [[#644](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/644)]
+- Bump IDE compatibility to support 2025.3 [[#642](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/642)]
 
 ## [0.0.49] - 2025-08-05
 
@@ -257,7 +267,8 @@
 - Add --exit-zero option to argument [[#12](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/12)]
 - Fix textRange logic [[#13](https://github.com/koxudaxi/ruff-pycharm-plugin/pull/13)]
 
-[Unreleased]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.49...HEAD
+[Unreleased]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.51...HEAD
+[0.0.51]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.49...v0.0.51
 [0.0.49]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.48...v0.0.49
 [0.0.48]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.47...v0.0.48
 [0.0.47]: https://github.com/koxudaxi/ruff-pycharm-plugin/compare/v0.0.46...v0.0.47